+++ CHANGELOG.md
@@ -29,7 +29,7 @@ We use the following categories for changes:
 - Make Jaeger Event queryable using name and tags [#1553]
 - Reset inverted labels cache on epoch change [#1561]
 - Error `column "exemplar_label_values" does not exist (SQLSTATE 42703)` on ingesting exemplars [#1574]
-- `/api/v1/labels` to respond with labels of authorized tenants only [#1577]
+- `/labels` & `/label/{name}/values` to respond with labels of authorized tenants only [#1577]
 
 ## [0.13.0] - 2022-07-20